refactor: split UI and editor helper types
