+++ src/CADShared/Cad/Editor/PromptOptionsEx.cs
@@ -1,4 +1,3 @@
-// ReSharper disable UnusedAutoPropertyAccessor.Global
 namespace Fs.Fox.Cad;
 
 /// <summary>
@@ -66,24 +65,4 @@ public static class PromptOptionsEx
             return;
         throw new KeywordException(e.Input.ToUpper());
     }
-}
-
-/// <summary>
-/// 关键字错误
-/// </summary>
-public class KeywordException : Exception
-{
-    /// <summary>
-    /// 关键字错误
-    /// </summary>
-    /// <param name="input">关键字</param>
-    public KeywordException(string input)
-    {
-        Input = input;
-    }
-
-    /// <summary>
-    /// 关键字
-    /// </summary>
-    public string Input { get; }
 }
