+++ src/CADShared/Cad/UI/StatusBar/PaneEx.cs
@@ -154,22 +154,3 @@ public static class PaneEx
         CadApp.StatusBar.Update();
     }
 }
-
-/// <summary>
-/// 托盘边距类型
-/// </summary>
-public enum PaneMarginType : byte
-{
-    /// <summary>
-    /// 无
-    /// </summary>
-    NONE,
-    /// <summary>
-    /// 小边距
-    /// </summary>
-    SMALL,
-    /// <summary>
-    /// 大边距
-    /// </summary>
-    LARGE
-}
