+++ src/CADShared/Cad/UI/StatusBar/PaneMarginType.cs
@@ -1,0 +1,20 @@
+namespace Fs.Fox.Cad;
+
+/// <summary>
+/// 托盘边距类型
+/// </summary>
+public enum PaneMarginType : byte
+{
+    /// <summary>
+    /// 无
+    /// </summary>
+    NONE,
+    /// <summary>
+    /// 小边距
+    /// </summary>
+    SMALL,
+    /// <summary>
+    /// 大边距
+    /// </summary>
+    LARGE
+}
